api/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.5
django-cors-headers==3.13.0
django-ninja==0.18.0
idna==3.3
marshmallow==2.21.0
Pillow==9.1.1
psycopg2-binary==2.9.3
pydantic==1.9.1
requests==2.27.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
text-unidecode==1.3
transbank-sdk==3.0.0
typing_extensions==4.2.0
urllib3==1.26.9
whitenoise==6.2.0
libpq-dev
gunicorn